messaging_app/chats/auth.py

from rest_framework_simplejwt.authentication import JWTAuthentication

class CustomJWTAuthentication(JWTAuthentication):
    def authenticate(self, request):
        # You can customize token retrieval or verification here
        return super().authenticate(request)